+++ src/JavaRoundtrip/main.cpp
@@ -44,7 +44,7 @@
 
 void remove_crlf(char * s)
 {
-    int len = strlen(s);
+    size_t len = strlen(s);
 
     if (len != 0) {
         if (s[len - 1] == '\n')
